solved issues
